SuspectPages ps1 was broken
SuspectPages.ps1 was not capturing data correct, nor was it inserting
into the table.  There is another fix that will be coming out that will
handle the update portion if same record already exists, but i wanted to
get this quick fix out there, so the function actually performs
correctly.  I have tested with positive results.