Push to master (#80)
* adding to notin comparison

* added check if Job already running to stop apparent failures

* SuspectPages had to update foreach to reference correct list of Servers

Updated foreach ($server in $servers) to foreach ($server in
$sqlservers)

* SuspectPages ps1 was broken

SuspectPages.ps1 was not capturing data correct, nor was it inserting
into the table.  There is another fix that will be coming out that will
handle the update portion if same record already exists, but i wanted to
get this quick fix out there, so the function actually performs
correctly.  I have tested with positive results.

* Whoops, forgot to fix the catch routine.

Forgot to fix the catch routine with the updated items.

* Revert "Whoops, forgot to fix the catch routine."

This reverts commit 8f35834b8b1d345cbe41874d5606640ee6424e3d.

* Script clean up

* Update per bug #65

Thanks @gbargsley for finding the bug and providing the solution!

* Added SSRS reports for Web and Express Editions

Duplicated sample SSRS reports and embedded datasets into each report.

* Bug Fix 69

Missing shared function write-log and code was out of order

* inital commit

Fixed spelling typo

* initial commit

Change Collation as this is too limiting and there are more than 30
characters in some of my database collation settings

* Escaping whitespace in job script file path

* Removing redundant log file creations and moving shared functions loading

* Adding InstanceName to SELECT to support named instances.

* Fixing updates and preventing duplicates

* Fixing updates and preventing duplicates

* data source

* 20%

* data source

* SSRS changes

* Master (#79)

* updated appveyor.yml

to bring in line with dev branch

* Fixed appveyor

* Removed duplicate key

* version update

* Updated version number to 0.2.54